refactor tests more
