+++ src/test/java/mocks/TestInstallService.java
@@ -1,0 +1,52 @@
+package mocks;
+
+import engine.FormplayerConfigEngine;
+import installers.FormplayerInstallerFactory;
+import org.commcare.util.engine.CommCareConfigEngine;
+import org.springframework.beans.factory.annotation.Autowired;
+import services.FormplayerStorageFactory;
+import services.InstallService;
+
+import java.io.File;
+import java.net.URL;
+
+/**
+ * Created by willpride on 12/7/16.
+ */
+public class TestInstallService implements InstallService {
+
+    @Autowired
+    FormplayerStorageFactory storageFactory;
+
+    @Autowired
+    FormplayerInstallerFactory formplayerInstallerFactory;
+
+    @Override
+    public CommCareConfigEngine configureApplication(String reference, String username, String dbPath) {
+        try {
+            File dbFolder = new File(storageFactory.getDatabasePath());
+            dbFolder.delete();
+            dbFolder.mkdirs();
+            FormplayerConfigEngine engine = new FormplayerConfigEngine(storageFactory,
+                    formplayerInstallerFactory);
+            String absolutePath = getTestResourcePath(reference);
+            engine.initFromArchive(absolutePath);
+            engine.initEnvironment();
+            return engine;
+        } catch (Exception e) {
+            e.printStackTrace();
+            throw new RuntimeException(e);
+        }
+    }
+
+    private String getTestResourcePath(String resourcePath){
+        try {
+            URL url = this.getClass().getClassLoader().getResource(resourcePath);
+            File file = new File(url.getPath());
+            return file.getAbsolutePath();
+        } catch(NullPointerException npe){
+            npe.printStackTrace();
+            throw npe;
+        }
+    }
+}
